Changed VerticalScrollView as widget
